rename folder name to models, since it's concerning database models
