+++ ldb/models/alumnus.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals
 from django.db import models
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import gettext as _
-from ldb.formTemplates.person import Person
+from ldb.models.person import Person
 
 CONTACT_METHOD_CHOICES = (
     ('m', 'Mail'),
