+++ ldb/models/entity.py
