+++ ldb/models/organization.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from django.utils.encoding import python_2_unicode_compatible
 
-from ldb.formTemplates.entity import Entity
+from ldb.models.entity import Entity
 from django.db import models
 from django.urls import reverse
 from django.utils.translation import gettext as _
