+++ ldb/search_indexes.py
@@ -2,8 +2,8 @@ from __future__ import unicode_literals
 
 from haystack import indexes
 
-from ldb.formTemplates.organization import Organization
-from ldb.formTemplates.person import Person
+from ldb.models.organization import Organization
+from ldb.models.person import Person
 
 
 class PersonIndex(indexes.SearchIndex, indexes.Indexable):
