+++ ldb/tests/test_membership_status.py
@@ -1,8 +1,8 @@
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 
-from ldb.formTemplates.membershipStatus import MembershipStatus
-from ldb.formTemplates.student import Student
+from ldb.models.membershipStatus import MembershipStatus
+from ldb.models.student import Student
 from ldb.tests.helpers import LDBHelperMixin
